+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
             android:name=".MapActivity"
             android:exported="false" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".settings.SettingsActivity"
             android:exported="false" />
         <activity
             android:name=".CameraActivity" />
